% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesVFL_Function.R
\name{BayesVFLReg}
\alias{BayesVFLReg}
\title{Runs MCMC for BayesVFLReg}
\usage{
BayesVFLReg(Phi_Y, Phi_X, nsamps, burnin = 500)
}
\arguments{
\item{Phi_Y}{An m x Q matrix containing the globally sketched response matrix. (elements must be numeric)}

\item{Phi_X}{A m x P matrix containing the globally sketched covariate matrix. (elements must be numeric)}

\item{nsamps}{Number of desired post burn-in samples.}

\item{burnin}{Number of burn-in iterations, default 500}
}
\value{
A named list containing the following components:
* **B**: An array of the posterior samples for the B matrix (nsamps x P x Q).
* **A**: An array of the posterior samples for the A matrix (nsamps x Q x Q).
* **sigma**: A matrix of posterior samples of sigma^2 (nsamps x Q).
* **lambda**: The posterior local shrinkage parameters.
* **tau**: The posterior global shrinkage parameters.
}
\description{
Takes in the globally sketched data and performs the MCMC for the High-Dimensional Sparse Reduced Rank Regression.
}
