% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_Phi.R
\name{gnerate_Phi}
\alias{gnerate_Phi}
\title{Generates Sketching matrix Phi}
\usage{
gnerate_Phi(n, m)
}
\arguments{
\item{n}{number of samples}

\item{m}{desired sketch size}
}
\value{
Phi: A m x n matrix.
}
\description{
Takes in n (number of samples) and m (desired sketch size) to generate Phi.
}
