% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{post_process}
\alias{post_process}
\title{Post processing}
\usage{
post_process(
  samps = NULL,
  B = NULL,
  A = NULL,
  B_est = NULL,
  A_est = NULL,
  C_est = NULL
)
}
\arguments{
\item{samps}{The output object of BayesVFLReg() of this package (default NULL).}

\item{B}{Posterior draws of the B matrix (nsamps x P x Q) (default NULL).}

\item{A}{Posterior draws of the A matrix (nsamps x Q x Q) (default NULL).}

\item{B_est}{Estimate of the B matrix (P x Q)  (default NULL).}

\item{A_est}{Estimate of the A matrix (Q x Q) (default NULL).}

\item{C_est}{Estimate of the C (coefficient matrix) (P x Q)  (default NULL).}
}
\value{
A vector of length P of TRUE indicating a significant coefficient.
}
\description{
Takes posterior draws to
}
